Allow NPM data to be updated for a single addition, in addition to updating in bulk
This moves the old npm-fetch/fetch.js script to npm-fetch/fetch-all.js. This still works exactly the same way.
The old "npm:update" Rake task is now "npm:update:all".
A new Rake task called "npm:update" takes an addon name as a paremeter and updates the data for that addon only.
The new npm-fetch/fetch.js script takes a package name as a command-line argument and outputs the package data (as JSON) to STDOUT.